requirements.txt
alembic==1.16.5
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
python-docx==1.2.0
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
